+++ Triangle.pde
@@ -4,7 +4,7 @@ class Triangle extends Shape{
   float x2;
   float y2;
   
-  Triangle(int x, int y, float x1, float y1, float x2, float y2){
+  Triangle(float x, float y, float x1, float y1, float x2, float y2){
     super(x,y);
     this.x1 = x1;
     this.y1 = y1;
@@ -12,7 +12,7 @@ class Triangle extends Shape{
     this.y2 = y2;
   }
   float calculateArea(){
-    return abs(x * (y1 - y2) + x1* (y2 - y) + x2 * (y - y1)) /2;
+    return abs(x * (y1 - y2) + x1 * (y2 - y) + x2 * (y - y1)) /2;
   }
   void display(){
     stroke(255, 0, 0);
